front: editor - load editor schema by using rtk query
